backend/LiveKit_Ustaad-Agent/requirements.txt
# LiveKit Agent dependencies
livekit-agents==1.2.8
livekit==1.0.12
livekit-api==1.0.5
livekit-protocol==1.0.5

# LiveKit plugins
livekit-plugins-openai==1.2.8
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-turn-detector==1.2.8
livekit-plugins-upliftai==1.2.8
livekit-plugins-silero==1.2.8

# Socket.IO TTS dependencies
python-socketio[asyncio]>=5.11.0
aiohttp>=3.9.0


python-dotenv>=1.0.0
loguru>=0.7.0

# OpenAI for STT/LLM
openai>=1.84.0
